Fixed the player model being disabled.
Updated the player to use Finlay's new temperature system.

Fixed bugs with the player movement like jumping up walls.

Fixed a bug in the GameManager's order that caused GetPlayer() to return null during scene startup when the player was actually there.

Updated the PlayerModel to have the particle effect instead of the Player object (fixes the particles rotating).

Added the PLAYER_TAG constant to the GameManager, use it to get the globally defined tag for the player.

Cleaned up the Player class and removed most of the debug code.